requirements_b.txt
backports.functools-lru-cache-1.5
beautifulsoup4-4.7.1
bs4-0.0.1
soupsieve-1.9
mysql-python-1.2.5
backports.csv-1.0.7
backports.functools-lru-cache-1.5
beautifulsoup4-4.7.1
certifi-2019.3.9
chardet-3.0.4
cheroot-6.5.4
cherrypy-17.4.1
contextlib2-0.5.5
feedparser-5.2.1
future-0.17.1
idna-2.8
jaraco.functools-2.0
lxml-4.3.3
more-itertools-5.0.0
mysqlclient-1.4.2.post1
nltk-3.4
numpy-1.16.2
pattern-3.6
pdfminer-20140328
portend-2.4
python-docx-0.8.10
pytz-2019.1
requests-2.21.0
scipy-1.2.1
setuptools-41.0.0
singledispatch-3.4.0.3
six-1.12.0
soupsieve-1.9
tempora-1.14
urllib3-1.24.1
zc.lockfile-1.4
textrazor-1.4.0
functools32-3.2.3.post2
SPARQLWrapper-1.8.2
isodate-0.6.0
pyparsing-2.4.0
rdflib-4.2.2
six-1.12.0
lru-0.1
xlsxwriter-1.1.6
py2neo
